update update data

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/dailyperform/breaktime/UpdateBreakTimeByTimeLeaveChangeHandler.java
@@ -10,6 +10,8 @@ import nts.arc.layer.app.command.CommandHandlerContext;
 import nts.arc.layer.app.command.CommandHandlerWithResult;
 import nts.arc.time.GeneralDate;
 import nts.uk.ctx.at.record.app.command.dailyperform.DailyCorrectEventServiceCenter;
+import nts.uk.ctx.at.record.app.command.dailyperform.DailyCorrectEventServiceCenter.EventHandleAction;
+import nts.uk.ctx.at.record.app.command.dailyperform.DailyCorrectEventServiceCenter.EventHandleResult;
 import nts.uk.ctx.at.record.app.find.dailyperform.resttime.dto.BreakTimeDailyDto;
 import nts.uk.ctx.at.record.dom.breakorgoout.BreakTimeOfDailyPerformance;
 import nts.uk.ctx.at.record.dom.breakorgoout.enums.BreakType;
@@ -30,7 +32,7 @@ import nts.uk.shr.com.context.AppContexts;
 
 /** Event：休憩時間帯を補正する */
 @Stateless
-public class UpdateBreakTimeByTimeLeaveChangeHandler extends CommandHandlerWithResult<UpdateBreakTimeByTimeLeaveChangeCommand, BreakTimeOfDailyPerformance> {
+public class UpdateBreakTimeByTimeLeaveChangeHandler extends CommandHandlerWithResult<UpdateBreakTimeByTimeLeaveChangeCommand, EventHandleResult<BreakTimeOfDailyPerformance>> {
 
 	@Inject
 	private WorkInformationRepository workInfoRepo;
@@ -52,34 +54,40 @@ public class UpdateBreakTimeByTimeLeaveChangeHandler extends CommandHandlerWithR
 
 	@Override
 	/** 休憩時間帯を補正する */
-	protected BreakTimeOfDailyPerformance handle(CommandHandlerContext<UpdateBreakTimeByTimeLeaveChangeCommand> context) {
+	protected EventHandleResult<BreakTimeOfDailyPerformance> handle(CommandHandlerContext<UpdateBreakTimeByTimeLeaveChangeCommand> context) {
 		UpdateBreakTimeByTimeLeaveChangeCommand command = context.getCommand();
 		WorkInfoOfDailyPerformance wi = command.cachedWorkInfo.orElse(getDefaultWorkInfo(command));
 		if(wi == null) {
-			return null;
+			return EventHandleResult.withResult(EventHandleAction.ABORT, null);
 		};
 		String companyId = command.companyId.orElse(AppContexts.user().companyId());
 		
 		WorkType wt = command.cachedWorkType.orElse(getDefaultWorkType(wi.getRecordInfo().getWorkTypeCode().v(), companyId));
 		if(wt == null) {
-			return null;
+			return EventHandleResult.withResult(EventHandleAction.ABORT, null);
 		}
 		
 		if (!wt.isWokingDay()) {
 			/** 「日別実績の休憩時間帯」を削除する */
-			this.breakTimeRepo.delete(command.getEmployeeId(), command.getWorkingDate());
-			return null;
+			if(!command.actionOnCache){
+				this.breakTimeRepo.delete(command.getEmployeeId(), command.getWorkingDate());
+			}
+			return EventHandleResult.withResult(EventHandleAction.DELETE, null);
 		}
 		
 		BreakTimeOfDailyPerformance breakTime = getUpdateBreakTime(command, wi, companyId);
 		if(breakTime != null){
 			/** 「日別実績の休憩時間帯」を更新する */
-			this.breakTimeRepo.update(breakTime);
-		} else {
-			/** 「日別実績の休憩時間帯」を削除する */
+			if(!command.actionOnCache){
+				this.breakTimeRepo.update(breakTime);
+			}
+			return EventHandleResult.withResult(EventHandleAction.UPDATE, breakTime);
+		}
+		/** 「日別実績の休憩時間帯」を削除する */
+		if(!command.actionOnCache){
 			this.breakTimeRepo.delete(command.getEmployeeId(), command.getWorkingDate());
 		}
-		return breakTime;
+		return EventHandleResult.withResult(EventHandleAction.DELETE, null);
 	}
 
 	/** 「補正した休憩時間帯」を取得する */
